Removed Image worker interface. Introduced instead new module to extract WSI ROIs.
