+++ dicoogle/pom.xml
@@ -466,5 +466,11 @@
             <artifactId>raven-log4j2</artifactId>
             <version>5.0.2</version>
         </dependency>
+        <dependency>
+            <groupId>org.dcm4che</groupId>
+            <artifactId>dcm4che-imageio</artifactId>
+            <version>3.3.7</version>
+            <scope>compile</scope>
+        </dependency>
     </dependencies>
 </project>
