+++ dicoogle/src/main/java/pt/ua/dicoogle/server/web/dicom/ROIExtractor.java
@@ -1,0 +1,2 @@
+package pt.ua.dicoogle.server.web.dicom;public class ROIExtractor {
+}
