+++ sdk/src/main/java/pt/ua/dicoogle/sdk/datastructs/wsi/WSIFrame.java
@@ -1,0 +1,2 @@
+package pt.ua.dicoogle.sdk.datastructs.wsi;public class WSIFrame {
+}
